+++ tests/testthat/test-read_sources.R
@@ -586,3 +586,9 @@ test_that("reading file with xmlMaps works", {
   expect_true(!is.null(wb$xmlMaps))
   expect_silent(wb$save("/tmp/test.xlsx"))
 })
+
+test_that("skipping builtins works", {
+  skip_online_checks()
+  xlsxFile <- testfile_path("Detailed_occupation_by_sex_and_education_ACS_2022_tab1.xlsx")
+  expect_equal(nrow(wb_load(xlsxFile)$get_named_regions(builtins = FALSE)), 0)
+})